requirements.txt
python-telegram-bot==13.15
openai==0.28.0